Start make detector_base a true base class. Renaming detector_sql to db_sql (because it is a db interface and not a detector class. Start implementing an actual sql detector class
