+++ NuRadioReco/detector/detector_base.py
@@ -33,97 +33,6 @@ class DateTimeSerializer(Serializer):
         return datetime.strptime(s, '%Y-%m-%dT%H:%M:%S')
 
 
-def buffer_db(in_memory, filename=None):
-    """
-    buffers the complete SQL database into a TinyDB object (either in memory or into a local JSON file)
-
-    Parameters
-    ----------
-    in_memory: bool
-        if True: the mysql database will be buffered as a tiny tb object that only exists in memory
-        if False: the mysql database will be buffered as a tiny tb object and saved in a local json file
-    filename: string
-        only relevant if `in_memory = True`: the filename of the json file of the tiny db object
-    """
-    serialization = SerializationMiddleware()
-    serialization.register_serializer(DateTimeSerializer(), 'TinyDate')
-    logger.info("buffering SQL database on-the-fly")
-    if in_memory:
-        db = TinyDB(storage=MemoryStorage)
-    else:
-        db = TinyDB(filename, storage=serialization, sort_keys=True, indent=4, separators=(',', ': '))
-    db.truncate()
-
-    from NuRadioReco.detector import detector_sql
-    sqldet = detector_sql.Detector()
-    results = sqldet.get_everything_stations()
-    table_stations = db.table('stations')
-    table_stations.truncate()
-    for result in results:
-        table_stations.insert({'station_id': result['st.station_id'],
-                               'commission_time': result['st.commission_time'],
-                               'decommission_time': result['st.decommission_time'],
-                               'station_type': result['st.station_type'],
-                               'position': result['st.position'],
-                               'board_number': result['st.board_number'],
-                               'MAC_address': result['st.MAC_address'],
-                               'MBED_type': result['st.MBED_type'],
-                               'pos_position': result['pos.position'],
-                               'pos_measurement_time': result['pos.measurement_time'],
-                               'pos_easting': result['pos.easting'],
-                               'pos_northing': result['pos.northing'],
-                               'pos_altitude': result['pos.altitude'],
-                               'pos_zone': result['pos.zone'],
-                               'pos_site': result['pos.site']})
-
-    table_channels = db.table('channels')
-    table_channels.truncate()
-    results = sqldet.get_everything_channels()
-    for channel in results:
-        table_channels.insert({'station_id': channel['st.station_id'],
-                               'channel_id': channel['ch.channel_id'],
-                               'commission_time': channel['ch.commission_time'],
-                               'decommission_time': channel['ch.decommission_time'],
-                               'ant_type': channel['ant.antenna_type'],
-                               'ant_orientation_phi': channel['ant.orientation_phi'],
-                               'ant_orientation_theta': channel['ant.orientation_theta'],
-                               'ant_rotation_phi': channel['ant.rotation_phi'],
-                               'ant_rotation_theta': channel['ant.rotation_theta'],
-                               'ant_position_x': channel['ant.position_x'],
-                               'ant_position_y': channel['ant.position_y'],
-                               'ant_position_z': channel['ant.position_z'],
-                               'ant_deployment_time': channel['ant.deployment_time'],
-                               'ant_comment': channel['ant.comment'],
-                               'cab_length': channel['cab.cable_length'],
-                               'cab_reference_measurement': channel['cab.reference_measurement'],
-                               'cab_time_delay': channel['cab.time_delay'],
-                               'cab_id': channel['cab.cable_id'],
-                               'cab_type': channel['cab.cable_type'],
-                               'amp_type': channel['amps.amp_type'],
-                               'amp_reference_measurement': channel['amps.reference_measurement'],
-                               'adc_id': channel['adcs.adc_id'],
-                               'adc_time_delay': channel['adcs.time_delay'],
-                               'adc_nbits': channel['adcs.nbits'],
-                               'adc_n_samples': channel['adcs.n_samples'],
-                               'adc_sampling_frequency': channel['adcs.sampling_frequency']})
-
-    results = sqldet.get_everything_positions()
-    table_positions = db.table('positions')
-    table_positions.truncate()
-    for result in results:
-        table_positions.insert({
-            'pos_position': result['pos.position'],
-            'pos_measurement_time': result['pos.measurement_time'],
-            'pos_easting': result['pos.easting'],
-            'pos_northing': result['pos.northing'],
-            'pos_altitude': result['pos.altitude'],
-            'pos_zone': result['pos.zone'],
-            'pos_site': result['pos.site']})
-
-    logger.info("sql database buffered")
-    return db
-
-
 @six.add_metaclass(NuRadioReco.utilities.metaclasses.Singleton)
 class DetectorBase(object):
     """
@@ -132,8 +41,7 @@ class DetectorBase(object):
     This class provides functions for all relevant detector properties.
     """
 
-    def __init__(self, source='json', json_filename='ARIANNA/arianna_detector_db.json',
-                 dictionary=None, assume_inf=True, antenna_by_depth=True):
+    def __init__(self, assume_inf=True, antenna_by_depth=True):
         """
         Initialize the stations detector properties.
         By default, a new detector instance is only created of none exists yet, otherwise the existing instance
@@ -143,47 +51,37 @@ class DetectorBase(object):
         
         Parameters
         ----------
-        source : str
-            'json', 'dictionary' or 'sql'
-            default value is 'json'
-            if dictionary is specified, the dictionary passed to __init__ is used
-            if 'sql' is specified, the file 'detector_sql_auth.json' file needs to be present in this folder that
-            specifies the sql server credentials (see 'detector_sql_auth.json.sample' for an example of the syntax)
-        json_filename : str
-            the path to the json detector description file (if first checks a path relative to this directory, then a
-            path relative to the current working directory of the user)
-            default value is 'ARIANNA/arianna_detector_db.json'
         assume_inf : Bool
             Default to True, if true forces antenna models to have infinite boundary conditions, otherwise the antenna madel will be determined by the station geometry.
         antenna_by_depth: bool (default True)
             if True the antenna model is determined automatically depending on the depth of the antenna. This is done by
             appending e.g. '_InfFirn' to the antenna model name.
             if False, the antenna model as specified in the database is used.
-        create_new: bool (default:False)
-            Can be used to force the creation of a new detector object. By default, the __init__ will only create a new
-            object if none already exists.
         """
         self._serialization = SerializationMiddleware()
         self._serialization.register_serializer(DateTimeSerializer(), 'TinyDate')
-        if source == 'sql':
-            self._db = buffer_db(in_memory=True)
-        elif source == 'dictionary':
-            self._db = TinyDB(storage=MemoryStorage)
-            self._db.truncate()
-            stations_table = self._db.table('stations', cache_size=1000)
-            for station in dictionary['stations'].values():
-                stations_table.insert(station)
-            channels_table = self._db.table('channels', cache_size=1000)
-            for channel in dictionary['channels'].values():
-                channels_table.insert(channel)
-        else:
-            self._db = TinyDB(
-                json_filename,
-                storage=self._serialization,
-                sort_keys=True,
-                indent=4,
-                separators=(',', ': ')
-            )
+        
+        self._db = self.buffer_detector_in_tiny_db()
+        
+        # if source == 'sql':
+        #     self._db = buffer_db(in_memory=True)
+        # elif source == 'dictionary':
+        #     self._db = TinyDB(storage=MemoryStorage)
+        #     self._db.truncate()
+        #     stations_table = self._db.table('stations', cache_size=1000)
+        #     for station in dictionary['stations'].values():
+        #         stations_table.insert(station)
+        #     channels_table = self._db.table('channels', cache_size=1000)
+        #     for channel in dictionary['channels'].values():
+        #         channels_table.insert(channel)
+        # else:
+        #     self._db = TinyDB(
+        #         json_filename,
+        #         storage=self._serialization,
+        #         sort_keys=True,
+        #         indent=4,
+        #         separators=(',', ': ')
+        #     )
 
         self._stations = self._db.table('stations', cache_size=1000)
         self._channels = self._db.table('channels', cache_size=1000)
@@ -205,8 +103,15 @@ class DetectorBase(object):
 
         self.__assume_inf = assume_inf
         if antenna_by_depth:
-            logger.info("the correct antenna model will be determined automatically based on the depth of the antenna")
+            logger.info("the correct antenna model will be determined automatically "
+                        "based on the depth of the antenna")
         self._antenna_by_depth = antenna_by_depth
+        
+        
+    def buffer_detector_in_tiny_db(self):
+        """ This class should be implemented in all derived classes """
+        raise NotImplementedError()
+
 
     def __query_channel(self, station_id, channel_id):
         Channel = Query()
